+++ crates/uv/tests/it/common/mod.rs
@@ -293,12 +293,25 @@ impl TestContext {
         self
     }
 
+    /// Adds a filter for platform-specific errors when a file is not executable.
+    #[inline]
+    pub fn with_filtered_not_executable(mut self) -> Self {
+        let pattern = if cfg!(unix) {
+            r"Permission denied \(os error 13\)"
+        } else {
+            r"\%1 is not a valid Win32 application. \(os error 193\)"
+        };
+        self.filters
+            .push((pattern.to_string(), "[PERMISSION DENIED]".to_string()));
+        self
+    }
+
     /// Adds a filter that ignores platform information in a Python installation key.
     pub fn with_filtered_python_keys(mut self) -> Self {
         // Filter platform keys
         let platform_re = r"(?x)
   (                         # We capture the group before the platform
-    (?:cpython|pypy)        # Python implementation
+    (?:cpython|pypy|graalpy)# Python implementation
     -
     \d+\.\d+                # Major and minor version
     (?:                     # The patch version is handled separately